third_party/blink/web_tests/external/wpt/css/cssom/stylesheet-cross-origin-redirect-quirks.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Cross-origin stylesheet with wrong MIME type
  assert_equals: cross-origin with wrong MIME type should fire error expected "error" but got "load"
[FAIL] Cross-origin redirect to same-origin with wrong MIME type
  assert_equals: cross-origin redirect to same-origin with wrong MIME type should fire error expected "error" but got "load"
[FAIL] Same-origin redirect to cross-origin with wrong MIME type
  assert_equals: same-origin redirect to cross-origin with wrong MIME type should fire error expected "error" but got "load"
[FAIL] Same-origin redirect to cross-origin redirect back to same-origin with wrong MIME type
  assert_equals: double redirect through cross-origin with wrong MIME type should fire error expected "error" but got "load"
Harness: the test ran to completion.
